requirements.txt
tensorflow==2.2.0
scikit-learn==0.20.4
numpy==1.16.6
opencv-python==4.2.0.32
grpcio==1.29.0
grpcio-tools==1.29.0
Pillow==6.2.2
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0